+++ packages/catalyst_builder_contracts/lib/src/annotation/generate_service_provider.dart
@@ -4,11 +4,6 @@ class GenerateServiceProvider {
   /// service provider.
   final String providerClassName;
 
-  /// Set this property to true if you like to include also the services
-  /// provided of your dependencies. Otherwise only the services in your
-  /// root package will be registered.
-  final bool includePackageDependencies;
-
   /// Mark this file as the entry point for the service container
   const GenerateServiceProvider({
     this.providerClassName = 'DefaultServiceProvider',